handle N/A values
