+++ src/styles/index.css
@@ -23,6 +23,7 @@
 ::after,
 ::before {
   box-sizing: border-box;
+  font-family: 'IBM Plex Sans', sans-serif;
 }
 
 .icon-person {
@@ -117,6 +118,11 @@
   bottom: auto;
   padding-left: 1.4em;
   transform: translate(0, 100%);
+}
+
+.icon-chart-disclaimer {
+  color: #777;
+  font-size: 10px;
 }
 
 .filters {
